minor fixes and docs
